[WIP] another major refactoring (rewrite) of solution handlers
I basically streamlined the whole concept of a solution and solution
data.

Now the basis is `StepSolutionData` representing the solution, error and
residual at a single time point (integration node).
To handle the trajectory of the solution over the course of multiple
time points, `TrajectorySolutionData` is an Iterable over
`StepSolutionData` objects. (Iterable because it provides __iter__ and
__next__ is provided by its return)

`ISolution` is the basic interface for all general solution objects
wrapping around a collection of `StepSolutionData` or
`TrajectorySolutionData` instances.
`FinalSolution` represents a solution object only storing the very last
time point of the very last iteration. Thus, internally, it only has a
single `StepSolutionData` object.
`FullSolution` provides an array of `TrajectorySolutionData` objects
representing the course of the solution trajectories over various
iterations.

ToDo:
- update test cases for the solution objects
- commit `.solvers.diagnosis.Error` and `.solvers.diagnosis.Residual`

Signed-off-by: Torbjörn Klatt <t.klatt@fz-juelich.de>