integrator: add Q-matrix + runtime fix
the integrator (read 'SDC integrator') now also computes and stores the
Q-matrix.
on integration it automagically uses the Q- (for full integral) or
S-matrix (for partial integrals).

runtime fix
the node type, weight function type and integrator types were passed as
actual instances before leading to instantiation on module load (which
is not adviced).
now only the class names are passed and instances are create internally
when needed.