support Python 3.4
Python 3.4 is raising errors in object.__format__(specifier) if it is
not overwritten by derived classes and the specifier is not empty.
Prior to 3.4 a DeprecationWarning was raised.

(cf. http://bugs.python.org/issue7994)

In all cases where this change affected PyPinT it's due to the fact,
that numpy.ndarray does not override object.__format__, thus there are
two possible solutions whenever we intent to stringify numpy.ndarrays:

- leave the format specifier empty: '{}'.format(my_ndarray)
- explicitly force to format as string: '{!s}'.format(my_ndarray)

I've decidet to use the first, as it's shorter and more flexible. It
lets numpy decide on the best way of formatting the array and possible
future changes in numpy to implement __format__ will be used directly
(as I understand).